requirements.txt
streamlit
st-pages
streamlit-extras
opencv-python-headless
pillow
